idam/requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
attrs==20.2.0
aws-sam-translator==1.27.0
aws-xray-sdk==2.6.0
bandit==1.6.2
black==20.8b1
boto==2.49.0
boto3==1.14.61
botocore==1.17.61
cachelib==0.1.1
certifi==2020.6.20
cffi==1.14.3
cfn-lint==0.37.1
chardet==3.0.4
click==7.1.2
coverage==5.3
cryptography==3.2
decorator==4.4.2
docker==4.3.1
docutils==0.15.2
ecdsa==0.14.1
envs==1.3
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.8
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsondiff==1.2.0
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.5.0
moto==1.3.16
mypy-extensions==0.4.3
networkx==2.6
packaging==20.4
pathspec==0.8.0
pbr==5.5.0
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycognito==0.1.4
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.4.2
pyjwt[crypto]
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.2
pytest-cov==2.10.1
pytest-env==0.6.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.2.0
pytz==2020.1
PyYAML==5.4
redis==3.5.3
regex==2020.7.14
requests==2.24.0
responses==0.12.0
rsa==4.7
s3transfer==0.3.3
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.19
sshpubkeys==3.1.0
stevedore==3.2.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.0
